Refactor buildings modal. Remove temp buildings. Connect cta.
